Alterando tipo da umidade de solo e atualizando string para o controle da lampada

+++ lib/components/card_plant.dart
@@ -17,7 +17,7 @@ class CardPlant extends StatefulWidget {
   final String apelidoPlanta;
   final String especiePlanta;
   final Function functionL;
-  double umidadeDoSolo;
+  int umidadeDoSolo;
   bool estadoLampada;
 
   @override
